bugfixing sesh
* you can no longer sell equipped runes
* reduce world berzerker power armor from 100/40 per level to 80/30 per level
* miniboss aura is back.
* sprees shouldn't display in invasion any more
* ctf auras for summonables should be blue or red depending on team
* blue shells on summonables and morphs are back on pvm modes
* reduced hp multiplier cap to 1.5 for cacodemons and medics
* brain corpses should now get removed
* prediction is no longer turned off with superspeed and sniper
* (lua) superspeed charge drain per frame reduced to 2/frame (from 4)
* 25% free vamp now only applies to brains and mutants. points in vampire will stack on top of that bonus for them.
* cacodemon should no longer play the pain animation when idling
* (lua) change flyer rocket delay to 1.5 seconds (from 3)
* (lua) change flyer hb damage from 35/4 to 40/5
* gi.soundindex and gi.sound were overridden with a custom implementation for gi.soundindex that uses an lru policy to replace sounds

+++ src/combat/abilities/caltrops.c
@@ -78,7 +78,7 @@ void ThrowCaltrops (edict_t *self, vec3_t start, vec3_t forward, int slevel, flo
 	self->client->ability_delay = level.time + CALTROPS_DELAY;
 	self->num_caltrops++;
 
-	safe_cprintf(self, PRINT_HIGH, "Caltrops deployed: %d/%d\n", self->num_caltrops, CALTROPS_MAX_COUNT);
+	safe_cprintf(self, PRINT_HIGH, "Caltrops deployed: %d/%d\n", self->num_caltrops, (int)CALTROPS_MAX_COUNT);
 
 	//  entity made a sound, used to alert monsters
 	self->lastsound = level.framenum;
@@ -98,9 +98,9 @@ void Cmd_Caltrops_f (edict_t *ent)
 	if (!V_CanUseAbilities(ent, CALTROPS, CALTROPS_COST, true))
 		return;
 
-	if (ent->num_caltrops >= CALTROPS_MAX_COUNT)
+	if (ent->num_caltrops >= (int)CALTROPS_MAX_COUNT)
 	{
-		safe_cprintf(ent, PRINT_HIGH, "You've reached the maximum number of caltrops (%d).\n", CALTROPS_MAX_COUNT);
+		safe_cprintf(ent, PRINT_HIGH, "You've reached the maximum number of caltrops (%d).\n", (int)CALTROPS_MAX_COUNT);
 		return;
 	}
